NEPT-279: Fix problems in login.
