+++ tests/src/Context/EcasContext.php
@@ -8,13 +8,16 @@ use Drupal\DrupalExtension\Context\RawDrupalContext;
  * Provides step definitions for interacting with ECAS.
  */
 class EcasContext extends RawDrupalContext {
+  use \Drupal\nexteuropa\Context\ContextUtil;
 
   /**
    * Creates and authenticates an ECAS user with the given role(s).
    *
+   * @see \Drupal\DrupalExtension\Context\DrupalContext::assertAuthenticatedByRole()
+   *
    * @Given I am logged in as an ECAS user with the :role role(s)
    */
-  public function assertAuthenticatedEcasByRole($role) {
+  public function ecasAssertAuthenticatedByRole($role) {
     // Check if a user with this role is already logged in.
     if (!$this->ecasLoggedInWithRole($role)) {
       // Create user (and project).
@@ -37,7 +40,7 @@ class EcasContext extends RawDrupalContext {
       }
 
       // Login.
-      $this->login();
+      $this->login($user);
     }
   }
 
@@ -47,16 +50,21 @@ class EcasContext extends RawDrupalContext {
    * @param string $role
    *   A single role, or multiple comma-separated roles in a single string.
    *
+   * @see RawDrupalContext::loggedInWithRole()
+   *
    * @return bool
    *   Returns TRUE if the current logged in user comes from
    *   Ecas and has this role (or roles).
    */
   public function ecasLoggedInWithRole($role) {
-    return $this->loggedIn() && $this->user && $this->user->ecas && isset($this->user->role) && $this->user->role == $role;
+    $user = $this->getUserManager()->getCurrentUser();
+    return $this->loggedIn() && $user && $user->ecas && isset($user->role) && $user->role == $role;
   }
 
   /**
    * Create an ECAS user.
+   *
+   * @see RawDrupalContext::userCreate()
    *
    * @return object
    *   The created user.
@@ -66,6 +74,7 @@ class EcasContext extends RawDrupalContext {
     $this->parseEntityFields('user', $user);
     $this->getDriver()->userCreate($user);
     $this->dispatchHooks('AfterUserCreateScope', $user);
+    $this->userManager->addUser($user);
 
     // Add the user to the authmap table.
     db_insert('authmap')
@@ -76,26 +85,31 @@ class EcasContext extends RawDrupalContext {
       ))
       ->execute();
     $user->ecas = 1;
-    $this->users[$user->name] = $this->user = $user;
+
     return $user;
   }
 
   /**
    * Remove any created ECAS users.
    *
+   * @see RawDrupalContext::cleanUsers()
+   *
    * @AfterScenario @Ecas
    */
-  public function cleanEcasUsers() {
+  public function ecasCleanUsers() {
     // Remove any users that were created.
-    if (!empty($this->users)) {
-      foreach ($this->users as $user) {
+    if ($this->userManager->hasUsers()) {
+      foreach ($this->userManager->getUsers() as $user) {
         $this->getDriver()->userDelete($user);
         db_delete('authmap')
           ->condition('uid', $user->uid)
           ->execute();
       }
       $this->getDriver()->processBatch();
-      $this->users = array();
+      $this->userManager->clearUsers();
+      if ($this->loggedIn()) {
+        $this->logout();
+      }
     }
   }
